+++ src/frontend/src/lib/stores/index.ts
@@ -1,0 +1,11 @@
+//groups all stores
+import { rootStore, regiStore } from './tasks/routing';
+import { authStore, actor, user } from './tasks/auth';
+
+export {
+  authStore,
+  actor,
+  user,
+  rootStore,
+  regiStore
+}
