+++ src/frontend/src/lib/stores/tasks/auth.ts
@@ -1,16 +1,11 @@
-import { Actor, HttpAgent, type ActorSubclass, type Identity } from '@dfinity/agent';
+import { Actor, HttpAgent, type Identity } from '@dfinity/agent';
 import { AuthClient } from '@dfinity/auth-client';
 import type { User, _SERVICE } from 'src/declarations/backend/backend.did';
 import { get, writable } from 'svelte/store';
-import { idlFactory } from '../../../../declarations/backend';
+import { idlFactory } from '../../../../../declarations/backend';
+import { AuthState, type BackendActor } from '../types';
 
-export enum AuthState {
-	LoggedOut,
-	LoggedIn,
-	Registered
-}
 
-type BackendActor = ActorSubclass<_SERVICE>;
 
 export const authStore = writable<AuthState>();
 export let actor = writable<BackendActor>();
