+++ src/frontend/src/lib/stores/tasks/index.ts
@@ -1,0 +1,10 @@
+//groups all effects/functions
+import { login, syncAuth, logout } from "./auth";
+
+
+export {
+  syncAuth,
+  login,
+  logout
+}
+
