+++ src/frontend/src/lib/stores/tasks/routing.ts
@@ -1,0 +1,6 @@
+import { writable } from 'svelte/store';
+import type { RootState, RegiState } from '../types';
+
+
+export const rootStore = writable<RootState>();
+export const regiStore = writable<RegiState>();
