+++ src/frontend/src/lib/stores/types.ts
@@ -1,5 +1,26 @@
+import type { ActorSubclass } from '@dfinity/agent';
+import type { _SERVICE } from 'src/declarations/backend/backend.did';
+
 //to declare all type structures
 
+//Backend based
+export enum AuthState {
+	LoggedOut,
+	LoggedIn,
+	Registered
+}
+export type BackendActor = ActorSubclass<_SERVICE>;
+//Frontend based
+export enum RegiState {
+	Username,
+	Profile,
+	Finished
+}
+export enum RootState {
+	Landing,
+	Register,
+	App
+}
 // export const ProfileSchema = z.object({
 //   created: z.bigint(),
 //   connect:
