+++ src/frontend/src/routes/(landing)/+page.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-	import { login } from '$lib/stores/auth';
+	import { login } from '$lib/stores/tasks';
 	import Proto1 from '$lib/assets/images/proto-1.svg?url';
 	import Segment from '$lib/components/landing/Segment.svelte';
 	import Grid from '$lib/components/landing/Grid.svelte';
