+++ src/frontend/src/routes/register/profile/+page.svelte
@@ -1,10 +1,8 @@
 <script lang="ts">
-	import { logout, actor, syncAuth } from '$lib/stores/auth';
 	import Button from '$lib/components/common/Button.svelte';
-	import { goto } from '$app/navigation';
 	import BorderBox from '$lib/components/common/BorderBox.svelte';
-	import { RegiState, regiStore } from '$lib/stores/stores';
-	import { z } from 'zod';
+	import { regiStore } from '$lib/stores';
+	import { RegiState } from '$lib/stores/types';
 
 	//import update profile call from +page.ts?
